First push, cleaning and deprecating unused files. Next PR: encapsulation
