+++ code/frontend/src/_tests_/ChatHistory.test.jsx
@@ -1,6 +1,5 @@
 /* eslint-disable no-undef */
 /* eslint-disable no-unused-vars */
-import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import ChatHistory from './ChatHistory';
 import { Context } from '../../context/ContextProvider';
