+++ code/frontend/src/_tests_/LoginPage.test.jsx
@@ -1,7 +1,6 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-undef */
 // LoginPage.test.js
-import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import LoginPage from '../LoginPage'; // Adjust the import path accordingly
