+++ code/frontend/src/_tests_/Sidebar.test.jsx
@@ -1,6 +1,5 @@
 /* eslint-disable no-undef */
 /* eslint-disable no-unused-vars */
-import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import Sidebar from './Sidebar';
 import { Context } from '../../context/ContextProvider';
