+++ code/frontend/src/assets/availableCourses.tsx
