+++ code/frontend/src/components/Sidebar/SaveButton.jsx
@@ -1,5 +1,4 @@
 /* eslint-disable no-unused-vars */
-import React, { useContext, useState } from 'react'
 import './Sidebar.css'
 import { assets } from '../../assets/assets'
 import { useChat } from '../../context/ChatContext'
