+++ code/frontend/src/components/Sidebar/depreciated_ShareButton.jsx
@@ -1,5 +1,4 @@
 /* eslint-disable no-unused-vars */
-import React, { useContext, useState } from 'react'
 import './Sidebar.css'
 import { assets } from '../../assets/assets'
