+++ code/frontend/src/components/depreciated_CourseProvider/deprecated_CourseProvider.jsx
