+++ code/frontend/src/pages/ChatPage.jsx
@@ -2,7 +2,6 @@
 // Created by Natt
 // Updated and annotated by Natasya Liew
 
-import React from 'react'; // Import React library
 import ChatBox from '../components/ChatBox/ChatBox'; // Import the ChatBox component for displaying chat messages
 import SideBar from '../components/Sidebar/Sidebar'; // Import the Sidebar component for navigation
 import { ChatProvider } from '../context/ChatContext'; // Import ChatProvider for managing chat state context
