+++ code/frontend/src/pages/SignupPage.jsx
@@ -3,7 +3,6 @@
 // Created by Natt 
 //Updated, integrated, and annotated by Natasya Liew
 
-import React, { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'; // // Import password visibility toggle icons
